Agregando Visor de carretilla y listado de los productos de la carretilla sin estilos
