+++ vistas/layout.tpl
@@ -13,6 +13,7 @@
                 <li><a href="index.php?page=libros">Libros</a></li>
                 <li><a href="index.php?page=postit">Postits</a></li>
                 <li><a href="index.php?page=productos">Catalogo</a></li>
+                <li><a href="index.php?page=carretilla">Carretilla <img src="publico/imagenes/carretillaflag.svg" style="vertical-align: middle; width:1.125em;padding:0;" /><b>{{ctdcarretilla}}</b></a></li>
     </ul>
     <div class="logo">
         <img src="publico/imagenes/logo.svg" />
